+++ packages/flame/CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Refer to the [CONTRIBUTING guide](https://github.com/lightspeed/flame/blob/master/.github/CONTRIBUTING.md) for more info.
 
+## [Unreleased]
+
+### Fixed
+
+- Destructure css and className to the wrapper component for Checkbox
+- Destructure css and className to the wrapper component for Radio
+
 ## 1.2.4 - 2019-11-22
 
 ### Deprecated